Move handlers into separata directory
